// src/app/atendimento-paciente/ficha-consulta-retorno/shared/model/avaliacao-composicao-corporal.model.ts
export class AvaliacaoComposicaoCorporal {

    id: number;
    pesoAtualKg: string;
	estaturaMetros: string;
	indiceMassaCorporalKgMetrosQuadrado: string;
	porcentagemGorduraDensidadeCorporal: string;
	dobraCutaneaTricipitalMilimetros: string;
	dobraCutaneaCoxaMilimetros: string;
	dobraCutaneaPanturrilhaMilimetros: string;
	dobraCutaneaPeitoralHomemMilimetros: string;
	dobraCutaneaAbdominalHomemMilimetros: string;
	dobraCutaneaSupraIliacaMulherMilimetros: string;
	resultadoCalculoDensidadeCorporal: string;
}